+++ aws-java-sdk-elasticsearch/src/main/java/com/amazonaws/services/elasticsearch/AbstractAWSElasticsearch.java
@@ -53,6 +53,11 @@ public class AbstractAWSElasticsearch implements AWSElasticsearch {
     }
 
     @Override
+    public DeleteElasticsearchServiceRoleResult deleteElasticsearchServiceRole(DeleteElasticsearchServiceRoleRequest request) {
+        throw new java.lang.UnsupportedOperationException();
+    }
+
+    @Override
     public DescribeElasticsearchDomainResult describeElasticsearchDomain(DescribeElasticsearchDomainRequest request) {
         throw new java.lang.UnsupportedOperationException();
     }