+++ aws-java-sdk-elasticsearch/src/main/java/com/amazonaws/services/elasticsearch/model/transform/ElasticsearchDomainConfigJsonUnmarshaller.java
@@ -68,6 +68,10 @@ public class ElasticsearchDomainConfigJsonUnmarshaller implements Unmarshaller<E
                     context.nextToken();
                     elasticsearchDomainConfig.setSnapshotOptions(SnapshotOptionsStatusJsonUnmarshaller.getInstance().unmarshall(context));
                 }
+                if (context.testExpression("VPCOptions", targetDepth)) {
+                    context.nextToken();
+                    elasticsearchDomainConfig.setVPCOptions(VPCDerivedInfoStatusJsonUnmarshaller.getInstance().unmarshall(context));
+                }
                 if (context.testExpression("AdvancedOptions", targetDepth)) {
                     context.nextToken();
                     elasticsearchDomainConfig.setAdvancedOptions(AdvancedOptionsStatusJsonUnmarshaller.getInstance().unmarshall(context));